+++ s/consts.ts
@@ -1,0 +1,9 @@
+
+import {deep} from "@e280/stz"
+
+export const consts = deep.freeze({
+	divisor: ".",
+	delimiter: ":",
+	chunkSize: 1024,
+})
+